fix issue that import the wrong file path of readFile
